+++ example/lib/home/home_controller.dart
@@ -18,6 +18,14 @@ class HomeController with AsyncStateMixin {
     }
   }
 
+  Future<void> loginWithAsyncLoaderHandler() async {
+    final loaderHandler = AsyncLoaderHandler()..start();
+    await Future.delayed(const Duration(seconds: 1));
+    await Future.delayed(const Duration(seconds: 5));
+
+    loaderHandler.close();
+  }
+
   ///Method that simulates a login error
   Future<void> loginError() async {
     try {